updated dataset directory to keep them within the repo

+++ src/download_syncan_dataset.sh
@@ -1,3 +1,5 @@
+#!/bin/bash
+
 # MIT License
 #
 # Copyright (c) 2023 Md Hasan Shahriar
@@ -25,14 +27,20 @@
 # Email: hshahriar@vt.edu
 #
 
+DIRECTORY="../datasets/can-ids/syncan/"
 # Downloading syncan dataset from the github repo
-git clone https://github.com/etas/SynCAN.git ../../datasets/can-ids/syncan/
-echo "Raw SynCAN dataset downloaded in ../../datasets/can-ids/syncan/"
-cd ../../datasets/can-ids/syncan/
-unzip 'train_*.zip' -d ambients
-echo "Unzipped training dataset in datasets/can-ids/syncan/ambients"
-unzip 'test_*.zip' -d attacks
-echo "Unzipped training dataset in datasets/can-ids/syncan/attacks"
-rm -rf *.zip
-rm -rf attacks/test_normal*
-echo "SyncCAN Data Downloaded!"
+if [ -d "$DIRECTORY" ]; then
+  echo "The folder $DIRECTORY already exists." 
+  echo "To update the dataset please delete $DIRECTORY and run this script again."
+else
+  git clone https://github.com/etas/SynCAN.git "$DIRECTORY"
+  echo "Raw SynCAN dataset downloaded in $DIRECTORY"
+  cd ../datasets/can-ids/syncan/
+  unzip 'train_*.zip' -d ambients
+  echo "Unzipped training dataset in datasets/can-ids/syncan/ambients"
+  unzip 'test_*.zip' -d attacks
+  echo "Unzipped training dataset in datasets/can-ids/syncan/attacks" 
+  rm -rf *.zip
+  rm -rf attacks/test_normal*
+  echo "SyncCAN Data Downloaded!"
+fi
